fix: correct the data types

+++ modules/Note/Database/Migrations/2020_02_05_182559_create_notes_table.php
@@ -18,14 +18,14 @@ class CreateNotesTable extends Migration
 
             // FK Relationships
             $table->unsignedBigInteger('org_id');
-            $table->integer('entity_type_id');
-            $table->integer('reference_id');
+            $table->unsignedBigInteger('entity_type_id');
+            $table->unsignedBigInteger('reference_id');
             $table->text('note')->nullable();
 
             //Audit Log Fields
             $table->unsignedBigInteger('created_by')->default(0);
             $table->unsignedBigInteger('updated_by')->nullable();
-            $table->integer('deleted_by')->nullable();
+            $table->unsignedBigInteger('deleted_by')->nullable();
 
             $table->timestamps();
             $table->softDeletes();
